requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langgraph==0.2.28
chromadb==0.5.5
streamlit==1.38.0
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.2
tiktoken==0.7.0
pypdf==4.3.1
python-docx==1.1.2
python-multipart==0.0.9
